chore: move DatabaseFixture to test, rename to TestFixture
